feat: 支持使用上下键切换输入历史
feat: 添加权限组页面
test: 新增测试用例

+++ src/Serein.Core/Models/Server/IServerInfo.cs
@@ -19,7 +19,7 @@ public interface IServerInfo : INotifyPropertyChanged
 
     public ulong InputLines { get; }
 
-    public double CPUUsage { get; }
+    public int CPUUsage { get; }
 
     public MineStat? Stat { get; }
 }
